move zipkinservice to server package
